+++ electrical/electrical/Source/Character/Chara_EnemyBase.cpp
@@ -1,7 +1,7 @@
 #include "Chara_EnemyBase.h"
 
-CharaEnemyBase::CharaEnemyBase(int x, int y, int radius,
-							   int speed, int hp, int graphHandle):
+CharaEnemyBase::CharaEnemyBase(float x, float y, int radius,
+							   float speed, int hp, int graphHandle):
 	CharaBase(x, y, radius, speed, hp, graphHandle)
 {
